add error handling to ot2 edit modules
