+++ protocol-designer/src/assets/localization/en/protocol_overview.json
@@ -2,7 +2,8 @@
   "add_gripper": "Add a gripper",
   "app_version": "{{version}} or higher",
   "author": "Organization/Author",
-  "conflict_on_slot": "Cannot move to this slot due to a module conflict",
+  "conflict_on_slot_labware": "Cannot move to this slot because the {{displayName}} is incompatible with this module",
+  "conflict_on_slot_module": "Cannot move to this slot due to a module conflict",
   "created": "Date created",
   "deck_hardware": "Deck hardware",
   "deck_slot": "Deck slot",
