+++ protocol-designer/src/components/molecules/MagnetModuleChangeContent/index.tsx
@@ -5,7 +5,7 @@ import {
   SPACING,
   StyledText,
 } from '@opentrons/components'
-import { KnowledgeLink } from '../../../components/organisms'
+import { KnowledgeLink } from '../../organisms'
 
 export function MagnetModuleChangeContent(): JSX.Element {
   const { t } = useTranslation('starting_deck_state')
