+++ protocol-designer/src/components/molecules/index.ts
@@ -7,6 +7,7 @@ export * from './HardwareOverlays'
 export * from './HotKeyDisplay'
 export * from './InputStepFormField'
 export * from './LiquidButton'
+export * from './MagnetModuleChangeContent'
 export * from './TextAreaField'
 export * from './ToggleExpandStepFormField'
 export * from './ToggleStepFormField'
