+++ protocol-designer/src/components/organisms/index.ts
@@ -22,6 +22,7 @@ export * from './LabwareUploadModal'
 export * from './LiquidsOverflowMenu'
 export * from './ModuleEmptySelectorButtons'
 export * from './Navigation'
+export * from './Ot2Modules'
 export * from './PipetteInfoItem'
 export * from './Portal'
 export * from './SelectPipetteModal'
