+++ protocol-designer/src/pages/Designer/DeckSetup/__tests__/MagnetModuleChangeContent.test.tsx
@@ -2,7 +2,7 @@ import { describe, it } from 'vitest'
 import { screen } from '@testing-library/react'
 import { i18n } from '../../../../assets/localization'
 import { renderWithProviders } from '../../../../__testing-utils__'
-import { MagnetModuleChangeContent } from '../MagnetModuleChangeContent'
+import { MagnetModuleChangeContent } from '../../../../components/molecules/MagnetModuleChangeContent'
 
 const render = () => {
   return renderWithProviders(<MagnetModuleChangeContent />, {
