+++ protocol-designer/src/pages/Hardware/index.tsx
@@ -16,7 +16,7 @@ import { getFileMetadata, getRobotType } from '../../file-data/selectors'
 import { getAdditionalEquipmentEntities } from '../../step-forms/selectors'
 import { useKitchen } from '../../components/organisms/Kitchen/hooks'
 import { FlexHardware } from './FlexHardware'
-import { Ot2Modules } from './Ot2Modules'
+import { Ot2Modules } from '../../components/organisms'
 import { NAV_BAR_HEIGHT_REM } from '../../components/atoms'
 
 export function Hardware(): JSX.Element {
