+++ protocol-designer/src/step-forms/actions/modules.ts
@@ -17,19 +17,6 @@ export const createModule = (
   type: 'CREATE_MODULE',
   payload: { ...args, id: `${uuid()}:${args.type}` },
 })
-export interface EditModuleAction {
-  type: 'EDIT_MODULE'
-  payload: {
-    id: string
-    model: ModuleModel
-  }
-}
-export const editModule = (
-  args: EditModuleAction['payload']
-): EditModuleAction => ({
-  type: 'EDIT_MODULE',
-  payload: args,
-})
 export interface DeleteModuleAction {
   type: 'DELETE_MODULE'
   payload: {
