+++ protocol-designer/src/tutorial/selectors.ts
@@ -67,19 +67,14 @@ export const shouldShowCoolingHint: Selector<boolean> = createSelector(
     return false
   }
 )
-export const shouldShowBatchEditHint: Selector<boolean> = createSelector(
-  getOrderedStepIds,
-  orderedStepIds => orderedStepIds.length >= 1
-)
 export const shouldShowWasteChuteHint: Selector<boolean> = createSelector(
   timelineFrameBeforeActiveItem,
   getUnsavedForm,
   getAdditionalEquipmentEntities,
   (prevTimelineFrame, unsavedForm, additionalEquipmentEntities) => {
-    const hasWasteChute =
-      Object.values(additionalEquipmentEntities).find(
-        ae => ae.name === 'wasteChute'
-      ) != null
+    const hasWasteChute = Object.values(additionalEquipmentEntities).some(
+      ae => ae.name === 'wasteChute'
+    )
     if (unsavedForm?.stepType !== 'moveLabware' || !hasWasteChute) {
       return false
     }
