+++ protocol-designer/src/types.ts
@@ -1,3 +1,9 @@
+import {
+  HEATERSHAKER_MODULE_TYPE,
+  MAGNETIC_MODULE_TYPE,
+  TEMPERATURE_MODULE_TYPE,
+  THERMOCYCLER_MODULE_TYPE,
+} from '@opentrons/shared-data'
 import type { FC } from 'react'
 import type { OutputSelector } from 'reselect'
 import type { NozzleConfigurationStyle } from '@opentrons/shared-data'
@@ -61,3 +67,9 @@ export interface RouteProps {
   path: string
   navLinkTo: string
 }
+
+export type OT2ModuleType =
+  | typeof MAGNETIC_MODULE_TYPE
+  | typeof TEMPERATURE_MODULE_TYPE
+  | typeof THERMOCYCLER_MODULE_TYPE
+  | typeof HEATERSHAKER_MODULE_TYPE
