+++ protocol-designer/src/ui/steps/actions/thunks/index.ts
@@ -276,7 +276,6 @@ export const saveStepForm: () => ThunkAction<any> = () => (
     )
     return
   }
-
   if (tutorialSelectors.shouldShowCoolingHint(initialState)) {
     dispatch(tutorialActions.addHint('thermocycler_lid_passive_cooling'))
   }
